fixed most of security warnings (#349)
* fixed most of security warnings

+++ src/board_controller/ironbci/ironbci.cpp
@@ -231,7 +231,7 @@ void IronBCI::read_thread ()
         // eeg
         for (unsigned int i = 0; i < eeg_channels.size (); i++)
         {
-            package[eeg_channels[i]] = eeg_scale * cast_24bit_to_int32 (b + 1 + 3 * i);
+            package[eeg_channels[i]] = (double)eeg_scale * cast_24bit_to_int32 (b + 1 + 3 * i);
         }
 
         package[board_descr["timestamp_channel"].get<int> ()] = get_timestamp ();
